refactor: extract crash report and update services from main class
Move crash report handling and update checking logic into separate service classes (CrashReportService and UpdateService) to improve code organization and maintainability. This reduces the size of the Main.java file and separates concerns.
